add rocketmq-streams-clients add new device message submit analyse method add mqtt to receive Datacollect service  standard message

+++ SmartCloud-app/src/main/java/www/bwsensing/com/device/common/facility/BwAngleData.java
@@ -81,7 +81,7 @@ public class BwAngleData {
         monitorData.setTimeStamp(timestamp);
         monitorData.setType("avg");
         monitorData.setDataId(dataId);
-        monitorData.setDataIdValue(value);
+        monitorData.setDataIdValue((double) value);
         return monitorData;
     }
 }
