+++ SmartCloud-domain/src/main/java/www/bwsensing/com/domain/device/model/data/MonitorData.java
@@ -14,22 +14,27 @@ public class MonitorData {
      * 分组编号
      */
     private Integer groupId;
+
     /**
      * 检测项
      */
     private String dataId;
+
     /**
      *数据值
      */
-    private Float dataIdValue;
+    private Double dataIdValue;
+
     /**
      *Sn编码
      */
     private String sn;
+
     /**
      *查询类型
      */
     private String type;
+
     /**
      *时间戳
      */
